+++ PythonAPI/carla/source/libcarla/Actor.cpp
@@ -161,6 +161,10 @@ void export_actor() {
       .def("set_light_state", &cc::Vehicle::SetLightState, (arg("light_state")))
       .def("set_wheel_steer_direction", &cc::Vehicle::SetWheelSteerDirection, (arg("wheel_location")), (arg("angle_in_deg")))
       .def("get_wheel_steer_angle", &cc::Vehicle::GetWheelSteerAngle, (arg("wheel_location")))
+      .def("set_wheel_pitch_angle", &cc::Vehicle::SetWheelPitchAngle, (arg("wheel_location")), (arg("angle_in_deg")))
+      .def("get_wheel_pitch_angle", &cc::Vehicle::GetWheelPitchAngle, (arg("wheel_location")))
+      .def("set_wheel_height", &cc::Vehicle::SetWheelHeight, (arg("wheel_location")), (arg("height")))
+      .def("get_wheel_height", &cc::Vehicle::GetWheelHeight, (arg("wheel_location")))
       .def("get_light_state", CONST_CALL_WITHOUT_GIL(cc::Vehicle, GetLightState))
       .def("apply_physics_control", &cc::Vehicle::ApplyPhysicsControl, (arg("physics_control")))
       .def("get_physics_control", CONST_CALL_WITHOUT_GIL(cc::Vehicle, GetPhysicsControl))
@@ -179,6 +183,7 @@ void export_actor() {
   class_<cc::Walker, bases<cc::Actor>, boost::noncopyable, boost::shared_ptr<cc::Walker>>("Walker", no_init)
       .def("apply_control", &ApplyControl<cr::WalkerControl>, (arg("control")))
       .def("apply_control", &ApplyControl<cr::WalkerBoneControl>, (arg("control")))
+      .def("get_bone_control_state", &cc::Walker::GetWalkerBoneState)
       .def("get_control", &cc::Walker::GetWalkerControl)
       .def(self_ns::str(self_ns::self))
   ;