+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
   * Added API functions to 3D vectors: `squared_length`, `length`, `make_unit_vector`, `dot`, `dot_2d`, `distance`, `distance_2d`, `distance_squared`, `distance_squared_2d`, `get_vector_angle`
   * Added API functions to 2D vectors: `squared_length`, `length`, `make_unit_vector`
+  * Make vehicle wheel pitch and height available through LibCarla.
+  * Make walker bone positions available through LibCarla.
 
 ## CARLA 0.9.12